requirements.txt
apex==0.9.10dev
azureml==0.2.7
boto3==1.26.74
botocore==1.29.74
capstone==4.0.2
Cython==0.29.33
deepspeed==0.8.1
editdistance==0.6.2
fairscale==0.4.13
fastBPE==0.1.0
gdown==4.6.3
gossip==2.4.0
hydra-core
iopath==0.1.10
matplotlib==3.6.1
networkx==2.8.7
nltk==3.8.1
numpy==1.23.3
omegaconf
psutil==5.9.3
pyarrow==9.0.0
pyelftools==0.29
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
sacrebleu==2.3.1
sacremoses==0.0.53
scikit_learn==1.2.1
scipy==1.10.0
sentencepiece==0.1.97
setuptools==63.4.1
soundfile==0.12.1
spacy==3.5.0
subword_nmt==0.3.8
tensorboardX==2.6
tokenizers==0.13.1
torch==1.12.1
torchaudio==0.12.1
torch_xla==1.0
tqdm==4.64.1
transformers==4.23.1
tweepy==4.12.1
unicorn==2.0.1.post1
wandb==0.13.4
